Update triangle_spec.rb
These edits fall more in line with the previous lesson where custom errors are explained. In all the examples in that lesson they nest the error class within the person class. In this lesson it is expected that it falls outside the Triangle class. This code better aligns with previous lessons.